add CREATE_PR_IF_NONE to mcpc-gen helper
